usb: move headers to include/usb/class
